build: improve build system to build doc on a separate branch
By creating and additional GitHub action configuration that triggers on every
tag, and updating the corrresponding project settings, now the contents
of GitHub pages should be taken from the action that deploys them.
This implied alo changing the git hooks of the project (that also were documented),
 updating the .gitignore to remove the generated docs, and removing some nps
commands that are now not needed.

fix #11