refactor: major refactor of configuration system
The major refactor is intended to provide a lazy loading of configuration, which allows to have
debugging information that loads properly when the -D or --debug is given through the CLI. When
using the API, the logger helper can be configured in order to have debugging information.

BREAKING CHANGE: The API of the configuration was completely changed. The name of the typedoc
configuration file, typedoc.js, was changed to typedoc.config.js to properly support windows
consoles.