feat: added new type of project - noCode
Now the user is available to select a "noCode" project type. This type of project may be used for
projects that do not require rollup, typescript and otherr tools, yet it still includes the nps
environment.
Additionally, improve over other projects and documentation in general as to be in ync with
the guidelines.