feat: full dependencie change
This commit bumps ESLint to v9 along with the ESlint configuration.
All configuration files had been updated as to use the .mjs or .cjs
extension, as to provide the proper style the file.
The tsconfig file also was updated, and the way in which we detect
if there is a local tsconfig in the local folder as to avoid generating one.
Styling had been changed, as to use @stylistic/eslint-plugin instead of
using prettier. Prettier is only used for non JS/TS files.

BREAKING CHANGE: The update for all configuration files is required for
all projects using the tool.