fix(cli): fix a bug preventing the use of package.json config when using CLI
When using the CLI, some values were set to default when not provided, thus preventing the correct
reading of the configuration in package.json. This prevented the use of some particular
configurations.

Additionally, re-adapt the prettier configuration to support using prettier in ESLint
and avoid using @stylistic, which brings more problems than solutions.